requirements.txt
# Core dependencies
setuptools>=60.0.0
absl-py==2.3.0
aiohappyeyeballs==2.6.1
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
python-binance>=1.0.19
binance-connector>=3.0.0
binance-futures-connector>=4.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
tenacity>=8.0.0
python-dateutil>=2.8.0
pandas>=2.0.0
numpy>=1.20.0
python-socketio>=5.0.0
eventlet>=0.33.0
Flask>=2.0.0
Flask-SocketIO>=5.0.0
requests>=2.25.0
websocket-client>=1.0.0
python-engineio>=4.0.0
python-socketio[client]>=5.0.0
ta-lib>=0.4.0
ta>=0.10.0
pytz>=2021.1
PyYAML>=6.0
python-decouple>=3.6
prometheus-client>=0.14.0
confluent-kafka>=2.0.0
tensorflow>=2.12.0
aiosignal==1.3.2
astunparse==1.6.3
attrs==25.3.0
bidict==0.23.1
binance-connector==3.12.0
binance-futures-connector==4.1.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
confluent-kafka==2.10.0
Cython==3.1.1
dateparser==1.2.1
dnspython==2.7.0
ta-lib==0.4.19
eventlet==0.40.1
Flask==3.1.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
frozenlist==1.6.2
gast==0.6.0
gevent==25.5.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.72.1
h11==0.16.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.4.4
namex==0.1.0
numpy==1.23.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
prometheus_client==0.21.1
propcache==0.3.1
protobuf==4.25.8
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
simple-websocket==1.1.0
six==1.17.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
